requirements.txt
python
django
postgreSql
pgAdmin4
psycopg2
djangorestframework
